Fix cast

+++ code/render/frame/framepass.cc
@@ -133,7 +133,7 @@ FramePass::Build(const BuildContext& ctx)
     for (IndexT i = 0; i < this->children.Size(); i++)
     {
         // Create new build context for each subpass
-        BuildContext newCtx = { myCompiled->pass, i, ctx.allocator, myCompiled->subpasses, ctx.events, ctx.barriers, ctx.buffers, ctx.textures };
+        BuildContext newCtx = { myCompiled->pass, (uint)i, ctx.allocator, myCompiled->subpasses, ctx.events, ctx.barriers, ctx.buffers, ctx.textures };
         this->children[i]->Build(newCtx);
     }
